requirements.txt
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
cryptography==3.1
decorator==4.4.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
Werkzeug==1.0.1
youtube-dl==2020.7.28
